requirements.txt
line-bot-sdk==3.11.0
openai==1.25.1